WIP - Remove hard coded Shifters and Timers
The code was hard coded to use Shifters 0 and 3 and timer 0,
but it could work just as easily with Shifters 4 and 7 or some combination
and any timer.

So I have it now checking which ones it can use...

Also commented out all of the places that reset the Flexio, as maybe not the only code trying to use a Flexio.

Also keep tract of which FlexIO logical mode (single, multiple, reset, read) we are in and if called again in the same state just return.